freerange: consolidate range contracts into option_ranges.ts
Contracts propagate within a file, and the call sites are constructors and
methods that freerange does not analyze wherever the contract lives, so one
module can hold them all and share the domain helpers across providers.

Add `pnpm fr` / `pnpm fr:audit`.
