+++ agents/package.json
@@ -35,6 +35,8 @@
     "clean:build": "pnpm clean && pnpm build",
     "lint": "eslint -f unix \"src/**/*.ts\"",
     "typecheck": "tsc -p tsconfig.typecheck.json",
+    "fr": "fr",
+    "fr:audit": "fr --audit",
     "api:check": "api-extractor run --typescript-compiler-folder ../node_modules/typescript",
     "api:update": "api-extractor run --local --typescript-compiler-folder ../node_modules/typescript --verbose",
     "throws:check": "throws-check src/**/*.ts"
