+++ agents/src/inference/stt.ts
@@ -8,6 +8,7 @@ import { APIError, APIStatusError } from '../_exceptions.js';
 import { AudioByteStream } from '../audio.js';
 import { type LanguageCode, areLanguagesEquivalent, normalizeLanguage } from '../language.js';
 import { log } from '../log.js';
+import { type RangedSTTModelOptions, checkSTTModelOptionRanges } from '../option_ranges.js';
 import { createStreamChannel } from '../stream/stream_channel.js';
 import {
   STT as BaseSTT,
@@ -211,220 +212,6 @@ export interface InworldSTTOptions {
   min_end_of_turn_silence_when_confident?: number;
   /** VAD threshold (0.0–1.0). Default: 0.5. */
   vad_threshold?: number;
-}
-
-/** Confidence and sensitivity knobs are probabilities: 0..1. */
-function checkedProbability(value: number): number {
-  console.assert(value >= 0);
-  console.assert(value <= 1);
-  return value;
-}
-
-/** Durations are never negative, whatever their unit. */
-function checkedDuration(value: number): number {
-  console.assert(value >= 0);
-  return value;
-}
-
-/** Cartesia turn-start threshold: 0.5-0.9. */
-function checkedTurnStartThreshold(value: number): number {
-  console.assert(value >= 0.5);
-  console.assert(value <= 0.9);
-  return value;
-}
-
-/** Cartesia eager turn-end threshold: 0.3-0.6. */
-function checkedTurnEagerEndThreshold(value: number): number {
-  console.assert(value >= 0.3);
-  console.assert(value <= 0.6);
-  return value;
-}
-
-/** Cartesia turn-end threshold: 0.05-0.5. */
-function checkedTurnEndThreshold(value: number): number {
-  console.assert(value >= 0.05);
-  console.assert(value <= 0.5);
-  return value;
-}
-
-/** Cartesia turn-end timeout: 640-11200 ms. */
-function checkedTurnEndTimeoutMs(value: number): number {
-  console.assert(value >= 640);
-  console.assert(value <= 11200);
-  return value;
-}
-
-/** Deepgram Flux eager end-of-turn threshold: 0.3-0.9. */
-function checkedEagerEotThreshold(value: number): number {
-  console.assert(value >= 0.3);
-  console.assert(value <= 0.9);
-  return value;
-}
-
-/** Deepgram Flux end-of-turn threshold: 0.5-0.9. */
-function checkedEotThreshold(value: number): number {
-  console.assert(value >= 0.5);
-  console.assert(value <= 0.9);
-  return value;
-}
-
-/** xAI endpointing: 0-5000 ms. */
-function checkedXaiEndpointingMs(value: number): number {
-  console.assert(value >= 0);
-  console.assert(value <= 5000);
-  return value;
-}
-
-/** Speechmatics max delay: 0.7-4.0 seconds. */
-function checkedSpeechmaticsMaxDelay(value: number): number {
-  console.assert(value >= 0.7);
-  console.assert(value <= 4);
-  return value;
-}
-
-/** Speaker counts are positive integers. */
-function checkedSpeakerCount(value: number): number {
-  console.assert(Number.isInteger(value));
-  console.assert(value >= 1);
-  return value;
-}
-
-/** Inworld voice-profile labels per category: 1-20. */
-function checkedVoiceProfileTopN(value: number): number {
-  console.assert(Number.isInteger(value));
-  console.assert(value >= 1);
-  console.assert(value <= 20);
-  return value;
-}
-
-/**
- * Caller requirements for the ranged provider option fields.
- *
- * Numeric fields are spelled out per provider rather than reusing the option
- * interfaces: freerange stops analyzing a function that reads a union-typed
- * field (`keyterm?: string | string[]`), and it does not analyze mapped types.
- */
-/** The ranged numeric fields across every provider's STT option interface. */
-type RangedSTTModelOptions = {
-  // Cartesia
-  turn_start_threshold?: number;
-  turn_eager_end_threshold?: number;
-  turn_end_threshold?: number;
-  turn_end_timeout_ms?: number;
-  min_volume?: number;
-  max_silence_duration_secs?: number;
-  // Deepgram, Deepgram Flux
-  endpointing?: number;
-  eager_eot_threshold?: number;
-  eot_threshold?: number;
-  eot_timeout_ms?: number;
-  // AssemblyAI, Inworld
-  end_of_turn_confidence_threshold?: number;
-  min_end_of_turn_silence_when_confident?: number;
-  max_turn_silence?: number;
-  voice_focus_threshold?: number;
-  inactivity_timeout_seconds?: number;
-  voice_profile_top_n?: number;
-  vad_threshold?: number;
-  // Speechmatics
-  max_delay?: number;
-  speaker_sensitivity?: number;
-  max_speakers?: number;
-  end_of_utterance_silence_trigger?: number;
-};
-
-function checkCartesiaOptionRanges(modelOptions: RangedSTTModelOptions): void {
-  if (modelOptions.turn_start_threshold !== undefined) {
-    checkedTurnStartThreshold(modelOptions.turn_start_threshold);
-  }
-  if (modelOptions.turn_eager_end_threshold !== undefined) {
-    checkedTurnEagerEndThreshold(modelOptions.turn_eager_end_threshold);
-  }
-  if (modelOptions.turn_end_threshold !== undefined) {
-    checkedTurnEndThreshold(modelOptions.turn_end_threshold);
-  }
-  if (modelOptions.turn_end_timeout_ms !== undefined) {
-    checkedTurnEndTimeoutMs(modelOptions.turn_end_timeout_ms);
-  }
-  if (modelOptions.min_volume !== undefined) checkedProbability(modelOptions.min_volume);
-  if (modelOptions.max_silence_duration_secs !== undefined) {
-    checkedDuration(modelOptions.max_silence_duration_secs);
-  }
-}
-
-function checkDeepgramFluxOptionRanges(modelOptions: RangedSTTModelOptions): void {
-  if (modelOptions.eager_eot_threshold !== undefined) {
-    checkedEagerEotThreshold(modelOptions.eager_eot_threshold);
-  }
-  if (modelOptions.eot_threshold !== undefined) checkedEotThreshold(modelOptions.eot_threshold);
-  if (modelOptions.eot_timeout_ms !== undefined) checkedDuration(modelOptions.eot_timeout_ms);
-}
-
-function checkDeepgramOptionRanges(modelOptions: RangedSTTModelOptions): void {
-  if (modelOptions.endpointing !== undefined) checkedDuration(modelOptions.endpointing);
-}
-
-function checkAssemblyAIOptionRanges(modelOptions: RangedSTTModelOptions): void {
-  if (modelOptions.end_of_turn_confidence_threshold !== undefined) {
-    checkedProbability(modelOptions.end_of_turn_confidence_threshold);
-  }
-  if (modelOptions.min_end_of_turn_silence_when_confident !== undefined) {
-    checkedDuration(modelOptions.min_end_of_turn_silence_when_confident);
-  }
-  if (modelOptions.max_turn_silence !== undefined) checkedDuration(modelOptions.max_turn_silence);
-  if (modelOptions.voice_focus_threshold !== undefined) {
-    checkedProbability(modelOptions.voice_focus_threshold);
-  }
-}
-
-function checkXaiOptionRanges(modelOptions: RangedSTTModelOptions): void {
-  if (modelOptions.endpointing !== undefined) checkedXaiEndpointingMs(modelOptions.endpointing);
-}
-
-function checkSpeechmaticsOptionRanges(modelOptions: RangedSTTModelOptions): void {
-  if (modelOptions.max_delay !== undefined) checkedSpeechmaticsMaxDelay(modelOptions.max_delay);
-  if (modelOptions.speaker_sensitivity !== undefined) {
-    checkedProbability(modelOptions.speaker_sensitivity);
-  }
-  if (modelOptions.max_speakers !== undefined) checkedSpeakerCount(modelOptions.max_speakers);
-  if (modelOptions.end_of_utterance_silence_trigger !== undefined) {
-    checkedDuration(modelOptions.end_of_utterance_silence_trigger);
-  }
-}
-
-function checkInworldSTTOptionRanges(modelOptions: RangedSTTModelOptions): void {
-  if (modelOptions.voice_profile_top_n !== undefined) {
-    checkedVoiceProfileTopN(modelOptions.voice_profile_top_n);
-  }
-  if (modelOptions.inactivity_timeout_seconds !== undefined) {
-    checkedDuration(modelOptions.inactivity_timeout_seconds);
-  }
-  if (modelOptions.end_of_turn_confidence_threshold !== undefined) {
-    checkedProbability(modelOptions.end_of_turn_confidence_threshold);
-  }
-  if (modelOptions.min_end_of_turn_silence_when_confident !== undefined) {
-    checkedDuration(modelOptions.min_end_of_turn_silence_when_confident);
-  }
-  if (modelOptions.vad_threshold !== undefined) checkedProbability(modelOptions.vad_threshold);
-}
-
-/**
- * Route the user's `modelOptions` to the option contracts of the provider that
- * will receive them. The routing itself is outside freerange's analyzed subset
- * (`String.startsWith`), so the per-provider functions above hold the contracts.
- */
-function checkSTTModelOptionRanges(
-  model: string | undefined,
-  modelOptions: RangedSTTModelOptions,
-): void {
-  if (model === undefined) return;
-  if (model.startsWith('cartesia/')) return checkCartesiaOptionRanges(modelOptions);
-  if (model.startsWith('deepgram/flux')) return checkDeepgramFluxOptionRanges(modelOptions);
-  if (model.startsWith('deepgram/')) return checkDeepgramOptionRanges(modelOptions);
-  if (model.startsWith('assemblyai/')) return checkAssemblyAIOptionRanges(modelOptions);
-  if (model.startsWith('xai/')) return checkXaiOptionRanges(modelOptions);
-  if (model.startsWith('speechmatics/')) return checkSpeechmaticsOptionRanges(modelOptions);
-  if (model.startsWith('inworld/')) return checkInworldSTTOptionRanges(modelOptions);
 }
 
 export type STTLanguages =
