+++ agents/src/voice/agent_session.ts
@@ -45,6 +45,7 @@ import { ToolContext, toToolContext } from '../llm/index.js';
 import type { LLMError } from '../llm/llm.js';
 import { log } from '../log.js';
 import { type ModelUsage, ModelUsageCollector, filterZeroValues } from '../metrics/model_usage.js';
+import { checkAgentSessionOptionRanges } from '../option_ranges.js';
 import { SimulationMode } from '../simulation.js';
 import type { STT } from '../stt/index.js';
 import type { STTError } from '../stt/stt.js';
@@ -195,43 +196,6 @@ export interface InternalSessionOptions<UserData> extends AgentSessionOptions<Us
   ttsReadIdleTimeout: number;
   forwardAudioIdleTimeout: number;
   ttsTextTransforms: readonly TextTransform[] | null;
-}
-
-/** Timeouts and durations are never negative. */
-function checkedDuration(value: number): number {
-  console.assert(value >= 0);
-  return value;
-}
-
-/** Step counts are non-negative integers. */
-function checkedCount(value: number): number {
-  console.assert(Number.isInteger(value));
-  console.assert(value >= 0);
-  return value;
-}
-
-/**
- * Caller requirements for the ranged {@link AgentSessionOptions} fields.
- *
- * Fields are spelled out rather than reusing the option type: freerange does
- * not analyze mapped types.
- */
-function checkAgentSessionOptionRanges(opts: {
-  maxToolSteps?: number;
-  userAwayTimeout?: number | null;
-  aecWarmupDuration?: number | null;
-  ttsReadIdleTimeout?: number;
-  forwardAudioIdleTimeout?: number;
-}): void {
-  if (opts.maxToolSteps !== undefined) checkedCount(opts.maxToolSteps);
-  if (opts.userAwayTimeout !== undefined && opts.userAwayTimeout !== null) {
-    checkedDuration(opts.userAwayTimeout);
-  }
-  if (opts.aecWarmupDuration !== undefined && opts.aecWarmupDuration !== null) {
-    checkedDuration(opts.aecWarmupDuration);
-  }
-  if (opts.ttsReadIdleTimeout !== undefined) checkedDuration(opts.ttsReadIdleTimeout);
-  if (opts.forwardAudioIdleTimeout !== undefined) checkedDuration(opts.forwardAudioIdleTimeout);
 }
 
 export const defaultAgentSessionOptions = {
